+++ client/webpack.pv.config.js
@@ -32,6 +32,11 @@ const config = {
                 ]
             }
         ]
+    },
+    optimization: {
+        splitChunks: {
+            chunks: 'all'
+        }
     }
 };